// src/lessons/Lesson_07/types.ts
export enum WEATHER_CODES {
  // SQ PO FC BR HZ FU DS SS - заранее типизированные
  SQ = "SQ",
  PO = "PA", // можно поменять один раз название в строке здесь в одном месте и всё
  FC = "FC",
  BR = "BR",
  HZ = "HZ",
  FU = "FU",
  DS = "DS",
  SS = "SS",
}